api/requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
marshmallow
marshmallow-SQLAlchemy
Flask-marshmallow
